+++ R/mutRate.R
@@ -3,15 +3,15 @@
 #' Calculate the overall mutation rate at a locus, given a mutation model an a
 #' set of allele frequencies.
 #'
-#' The mutation rate is found by the formula `1 - sum(diag(mut) * afreq)`.
+#' The mutation rate is found by the formula `1 - sum(diag(mutmat) * afreq)`.
 #'
-#' If `mut` is a [mutationModel()], the rate is calculated separately for the
-#' male and female matrices.
+#' If `mutmat` is a full [mutationModel()], the rate is calculated separately for
+#' the male and female matrices.
 #'
-#' @param mut A [mutationModel()] or [mutationMatrix()].
+#' @param mutmat A [mutationMatrix()] or [mutationModel()].
 #' @param afreq A vector of allele frequencies.
 #'
-#' @return A single number, or (if `mut` is a [mutationModel()] and the female
+#' @return A single number, or (if `mutmat` is a [mutationModel()] and the female
 #'   and male rates differ) a list of two numbers, named "female" and "male".
 #'
 #' @examples
@@ -21,16 +21,14 @@
 #'
 #' stopifnot(all.equal(r, 0.01))
 #' @export
-mutRate = function(mut, afreq = NULL) {
-  if(isMutationModel(mut)) {
-    r = lapply(mut, function(m) mutRate(m, afreq))
-
-    return(if(all.equal(r[1], r[2])) r[[1]] else r)
+mutRate = function(mutmat, afreq = NULL) {
+  if(isMutationModel(mutmat)) {
+    r = lapply(mutmat, function(m) mutRate(m, afreq))
+    return(if(all.equal(r$male, r$female)) r$male else unlist(r))
   }
 
-  if(is.null(afreq))
-    afreq = attr(mut, "afreq") %||%
+  afreq = afreq %||% attr(mutmat, "afreq") %||%
       stop2("Argument `afreq` is missing and not present as model attribute")
 
-  1 - sum(diag(mut) * afreq)
+  1 - sum(diag(mutmat) * afreq)
 }