+++ tests/testthat/test-mutationMatrix.R
@@ -1,7 +1,7 @@
 
-test_that("bad inputs are caught", {
+test_that("bad custom matrices are caught", {
   expect_error(mutationMatrix(),
-               'The "custom" model requires the argument `matrix` to be non-NULL')
+               "`matrix` cannot be NULL with the `custom` model")
   expect_error(mutationMatrix(matrix=data.frame(a=1)),
                "Custom matrix must be a matrix, not a data.frame")
   expect_error(mutationMatrix(matrix=list(a=1)),
@@ -16,26 +16,34 @@ test_that("bad inputs are caught", {
                "Length of `alleles` must equal the dimension of `matrix`")
   expect_error(mutationMatrix(matrix=matrix(1, ncol=1)),
                "When custom matrix lacks names, the argument `alleles` cannot be NULL")
+  m = matrix(1/2, 2, 2, dimnames = list(c("a", "b"), NULL))
+  expect_error(mutationMatrix(matrix = m),
+               "Mutation matrix has unequal row and column names")
+})
+
+test_that("bad inputs are caught", {
   expect_error(mutationMatrix(model="eq", matrix=matrix(1, ncol=1)),
-               "The `matrix` argument must be NULL when this model is specified")
+               "`matrix` cannot be used with the `equal` model")
   expect_error(mutationMatrix(model="eq"),
-               "`alleles` cannot be NULL with this model")
+               "`alleles` cannot be NULL with the `equal` model")
   expect_error(mutationMatrix(model="eq", alleles=1:2),
-               "`rate` cannot be NULL with this model")
+               "`rate` cannot be NULL with the `equal` model")
   expect_error(mutationMatrix(model="prop", alleles=1:2, rate=0),
-               "`afreq` cannot be NULL with this model")
+               "`afreq` cannot be NULL with the `proportional` model")
   expect_error(mutationMatrix(model="prop", alleles=1:2, afreq=1),
                "Frequency vector does not match the number of alleles")
   expect_error(mutationMatrix(model="prop", alleles=1:2, afreq=c(0.5, 0.501)),
                "Allele frequencies do not sum to 1")
   expect_error(mutationMatrix(model="prop", alleles=1:2, afreq=c(.5,.5), rate=2),
-               "Impossible mutation matrix; try reducing `rate`")
-  expect_error(mutationMatrix(model="step", alleles=1:2, rate=1, rate2=0.5, range=1),
+               "`rate` must be a number in the interval")
+  expect_error(mutationMatrix(model="prop", alleles=1:2, afreq=c(.3,.7), rate=1),
+               "Model undefined: max `rate` for the given input is: 0.6")
+  expect_error(mutationMatrix(model="step", alleles=1:2, rate=1, rate2=0.5, range=0.1),
                "The total mutation rate")
   expect_error(mutationMatrix(model="step", alleles=1:2, rate=0, rate2=0),
                "`range` cannot be NULL with the `stepwise` model")
-  expect_error(mutationMatrix(model="step", alleles=1:2, rate=0, rate2=0, range=0),
-               "`range` must be a positive number: 0")
+  expect_error(mutationMatrix(model="dawid", alleles=1:2, afreq = c(.3,.7), rate=0, rate2=0, range=1),
+               "For the `dawid` model, `range` must be in the interval")
 })
 
 
@@ -85,7 +93,7 @@ test_that("trivial model works", {
 
 test_that("trivial stepwise model is diagonal", {
   m = mutationMatrix(alleles = 1:3, model = "step",
-                     rate=0, rate2=0, range=99)
+                     rate=0, rate2=0, range=0.5)
 
   expect_silent(validateMutationMatrix(m, alleles = 1:3))
   expect_equivalent(m, mutationMatrix(matrix = diag(3), alleles = 1:3))