+++ NEWS.md
@@ -1,3 +1,30 @@
+# pedmut 0.8.0
+
+## New features
+
+* New function `makeReversible()` implementing three transformations to reversibility.
+
+* `mutationMatrix()` and `mutationModel()` gain a `transform` argument for applying transformations to reversibility on the fly.
+
+* New function `adjustRate()` for adjusting the overall mutation rate of a mutation matrix.
+
+* `mutationModel(model = "dawid", ...)` replaces the function `stepwiseReversible()`.
+
+* Random mutation models can now condition on a fixed overall mutation rate.
+
+* Include info on boundedness in the print method for mutation models.
+
+## Other
+
+* Refactored `mutationMatrix()` for better code maintainability.
+
+* Bug fix: Don't trivialise model name when there is only 1 allele.
+
+* Several minor improvements in documentation and examples.
+
+* Use rhub v2.
+
+
 # pedmut 0.7.0
 
 ## New features
@@ -40,7 +67,7 @@
 
 ## New features
 
-*  Various speedups, mainly concerning lumped models. These improvements should give better performance in other parts of the `ped suite`, especially in likelihood calculations with many markers.
+*  Various speedups, mainly concerning lumped models. These improvements should give better performance in other parts of the `pedsuite`, especially in likelihood calculations with many markers.
 
 * New function `lumpedModel()` which is a convenient wrapper of `lumpedMatrix()`.